pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "swiftbots"
version = "0.1.0"
description = "Python asynchronous web framework for rapid any-level bots development"
readme = "docs/README.md"
requires-python = ">=3.9.0"
dependencies = [
    "httpx>=0.0.0",
]

[project.optional-dependencies]
standard = []

grpc = [
    "grpcio>=0.0.0",
]

sqlalchemy = [
    "sqlalchemy>=0.0.0"
]

alembic = [
    "alembic>=0.0.0"
]

all = [
    "grpcio>=0.0.0",
    "sqlalchemy>=0.0.0",
    "alembic>=0.0.0"
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "black>=25.1.0",
    "grpcio-tools>=1.71.0",
    "mypy>=1.15.0",
    "pytest>=8.2.2",
    "pytest-timeout>=2.3.1",
    "ruff>=0.11.2",
]
